refactor(ui): domain 변경사항 적용
